feat(expiry_shortener): enforce maximum expiry duration
also changes the round trip time to be a u32

+++ crates/interledger/src/main.rs
@@ -327,7 +327,7 @@ pub fn main() {
                         send_routes: matches.is_present("send_routes"),
                         receive_routes: matches.is_present("receive_routes"),
                         routing_relation: value_t!(matches, "routing_relation", String).ok(),
-                        round_trip_time: value_t!(matches, "round_trip_time", u64).ok(),
+                        round_trip_time: value_t!(matches, "round_trip_time", u32).ok(),
                         packets_per_minute_limit: value_t!(
                             matches,
                             "packets_per_minute_limit",
